refactor(domain): 금액에 따른 로또 개수 계산을 Lottos 클래스에서 수행하도록 수정
- LottoService 를 삭제하고 Lottos에서 수행하도록 하여 클래스 구조 개선
