--- src/main/java/lotto/domain/LottoService.java
@@ -1,13 +1,0 @@
-package lotto.domain;
-
-public class LottoService {
-
-    public Lottos issueLottos(long amount) {
-        long count = calculateLottoCount(amount);
-        return Lottos.generateLottos(count);
-    }
-
-    private long calculateLottoCount(long amount) {
-        return amount / Lotto.LOTTO_PRICE;
-    }
-}
