+++ src/main/java/lotto/domain/Lottos.java
@@ -11,15 +11,26 @@ public class Lottos {
         this.lottos = List.copyOf(generatedLottos);
     }
 
-    public static Lottos generateLottos(long count) {
+    public static Lottos generateLottos(long amount) {
         List<Lotto> generatedLottos = new ArrayList<>();
+        long count = calculateLottoCount(amount);
         while (count-- > 0) {
             generatedLottos.add(new Lotto(LottoNumberGenerator.generate()));
         }
         return new Lottos(generatedLottos);
     }
 
-    public List<Lotto> getValue() {
+    private static long calculateLottoCount(long amount) {
+        return amount / Lotto.LOTTO_PRICE;
+    }
+
+    public List<Lotto> asList() {
         return lottos;
+    }
+
+    public long getTotalWinningAmount(WinningNumber winningNumber) {
+        return lottos.stream()
+                .mapToLong(lotto -> lotto.getWinningResult(winningNumber).getWinningAmount())
+                .sum();
     }
 }
